Initialize repository
I updated the original code to Rails 4.2.10. The original repo is available at
https://github.com/diogowernik/x_editable_rails_demo.git.

The gem x-editable-rails resides at https://github.com/werein/x-editable-rails.

I use diogowernik's repo because it supports the x-editable in-place editing
feature perfectly, without the headaching bug that the page has to be reloaded
to enable the in-place editing again.